fixed shadcn/ui linting issues, refactor theme-provider, revert minor change on view page
